refactor: model과 controller의 분리를 위한 리팩토링
