+++ src/main/java/baseball/Application.java
@@ -1,6 +1,6 @@
 package baseball;
 
-import nextstep.utils.Console;
+import controller.Game;
 
 public class Application {
     public static void main(String[] args) {
